+++ include/vb/util/Stream.h
@@ -6,12 +6,12 @@ namespace vb {
     template <typename T> using Stream = boost::coroutines2::detail::pull_coroutine<T>;
     template <typename T> using Sink   = boost::coroutines2::detail::push_coroutine<T>;
 
-    Stream<std::vector<int>> partitions(int n, int m = 1); // Partitions of n with piece size at least m
-    Stream<std::vector<int>> tuples(int k, int n);         // Ordered, distinct k-tuples in [0,n-1]
-    Stream<std::vector<int>> cycles(int k, int n);         // k-tuples up to cyclic permutation
+    Stream<std::vector<size_t>> partitions(size_t n, size_t m = 1); // Partitions of n with piece size at least m
+    Stream<std::vector<size_t>> tuples(size_t k, size_t n);         // Ordered, distinct k-tuples in [0,n-1]
+    Stream<std::vector<size_t>> cycles(size_t k, size_t n);         // k-tuples up to cyclic permutation
 
-    template <typename T> int size(Stream<T> &&S) {
-        int out = 0;
+    template <typename T> size_t size(Stream<T> &&S) {
+        size_t out = 0;
         for (const auto &i : S) {
             (void)i;
             out++;
@@ -19,7 +19,7 @@ namespace vb {
         return out;
     }
 
-    template <typename T> auto take(int n, Stream<T> &&S) {
+    template <typename T> auto take(size_t n, Stream<T> &&S) {
         return Stream<T>([n, S = std::move(S)](Sink<T> &yield) mutable {
             if (n <= 0) return;
             for (const auto &x : S) {
