+++ libvb/Stream.cpp
@@ -2,11 +2,11 @@
 #include <vector>
 
 namespace vb {
-    Stream<std::vector<int>> partitions(int n, int m) {
-        return Stream<std::vector<int>>([n, m](Sink<std::vector<int>> &yield) {
-            for (int i = m; i < n; ++i) {
+    Stream<std::vector<size_t>> partitions(size_t n, size_t m) {
+        return Stream<std::vector<size_t>>([n, m](Sink<std::vector<size_t>> &yield) {
+            for (size_t i = m; i < n; ++i) {
                 for (const auto &p : partitions(n - i, i)) {
-                    std::vector<int> out({i});
+                    std::vector<size_t> out({i});
                     for (auto j : p) out.push_back(j);
                     yield(out);
                 }
@@ -15,15 +15,15 @@ namespace vb {
         });
     }
 
-    Stream<std::vector<int>> tuples(int k, int n) {
-        return Stream<std::vector<int>>([k, n](Sink<std::vector<int>> &yield) {
+    Stream<std::vector<size_t>> tuples(size_t k, size_t n) {
+        return Stream<std::vector<size_t>>([k, n](Sink<std::vector<size_t>> &yield) {
             if (k == 0) {
                 yield({});
                 return;
             }
-            for (int i = 0; i < n; ++i) {
+            for (size_t i = 0; i < n; ++i) {
                 for (const auto &c : tuples(k - 1, n - 1)) {
-                    std::vector<int> out({i});
+                    std::vector<size_t> out({i});
                     for (auto j : c) out.push_back(j < i ? j : j + 1);
                     yield(out);
                 }
@@ -31,11 +31,11 @@ namespace vb {
         });
     }
 
-    Stream<std::vector<int>> cycles(int k, int n) {
-        return Stream<std::vector<int>>([k, n](Sink<std::vector<int>> &yield) {
-            for (int i = 0; i < n - k + 1; ++i) {
+    Stream<std::vector<size_t>> cycles(size_t k, size_t n) {
+        return Stream<std::vector<size_t>>([k, n](Sink<std::vector<size_t>> &yield) {
+            for (size_t i = 0; i < n - k + 1; ++i) {
                 for (const auto &c : tuples(k - 1, n - i - 1)) {
-                    std::vector<int> out(1, i);
+                    std::vector<size_t> out(1, i);
                     for (auto j : c) out.push_back(i + j + 1);
                     yield(out);
                 }
